Remove repetitive query of ClientConfig and CurrentWorkspace member
