+++ packages/twenty-front/src/modules/auth/states/isCurrentUserLoadingState.ts
@@ -1,0 +1,6 @@
+import { createState } from 'twenty-ui';
+
+export const isCurrentUserLoadedState = createState<boolean>({
+  key: 'isCurrentUserLoadedState',
+  defaultValue: false,
+});
