+++ packages/twenty-front/src/modules/client-config/states/isClientConfigLoadedState.ts
@@ -1,0 +1,6 @@
+import { createState } from 'twenty-ui';
+
+export const isClientConfigLoadedState = createState<boolean>({
+  key: 'isClientConfigLoadedState',
+  defaultValue: false,
+});
